Unbroke business registration, refactors
