+++ client/src/services/customerregistration.js
@@ -1,0 +1,8 @@
+import axios from '../utils/apiClient'
+
+const registerCustomer = async customerdata => {
+    const response = await axios.post('api/users/customers', customerdata)
+    return response.data
+}
+
+export default { registerCustomer }
